Refactor dashboard structure for better cohesion.
