+++ src/components/Dashboard/DashboardContent.js
@@ -1,10 +1,10 @@
 import React, {Component} from "react";
 import styled from "styled-components";
 import {Route, Switch} from "react-router-dom";
-import UploadFiles from "../UploadFiles/UploadFiles";
-import ExternalLink from "../ExternalLink/ExternalLink";
-import KibanaDashboard from "../KibanaDashboard/KibanaDashboard";
-import DashboardInfo from "../DashboardInfo/DashboardInfo";
+import UploadFiles from "./UploadFiles/UploadFiles";
+import ExternalLink from "./ExternalLink/ExternalLink";
+import KibanaDashboard from "./KibanaDashboard/KibanaDashboard";
+import DashboardInfo from "./DashboardInfo/DashboardInfo";
 
 
 const ContentContainer = styled.div`
