+++ src/components/Dashboard/KibanaDashboard/KibanaDashboard.js
